--- src/app/store/interfaces/cart-item.ts
@@ -1,9 +1,0 @@
-export interface ICartItem {
-  id: string;
-  product_id: number,
-  name: string,
-  img_url: string,
-  unit_price: number;
-  quantity: number;
-  ammount: number;
-}
